requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
django==5.0.6
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.1
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sqlparse==0.5.0
    # via django
tzdata==2024.1
    # via django
